restructure tools.py into  agent/tools folder
