+++ agent/src/tools/__init__.py
@@ -1,0 +1,2 @@
+from .base import register_tools
+__all__ = ["register_tools"]
